% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{Gibbs_mix}
\alias{Gibbs_mix}
\title{Gibbs_mix}
\usage{
Gibbs_mix(mu1, mu2, sigma1, sigma2, p, N, rv)
}
\arguments{
\item{mu1}{the mean of Y1}

\item{mu2}{the mean of Y2}

\item{sigma1}{the EM-estimated variance of Y1}

\item{sigma2}{the EM-estimated variance of Y2}

\item{p}{the probability of the mixture distribution}

\item{N}{the number of samples}

\item{rv}{the random number from mixed normal distribution}
}
\value{
the maximum likelihood estimation of the mean, variance and the mixed probablity
}
\description{
Gibbs sampling for two-component Gaussian Mixture
}
\examples{
\dontrun{
p<-0.6
N<-1000
set.seed(2335)
rv<-rv_mix(N,p)
mu1<-0.5
mu2<-1.5
sigma1<-sigma2<-sd(rv)
EM<-EM_mix(mu1,sigma1,mu2,sigma2,p)
sigma1<-EM[2]
sigma2<-EM[4]
p<-EM[5]
Gibbs_mix(mu1,mu2,sigma1,sigma2,p,N,rv)
}
}
