% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp for a bivariate density}
\usage{
gibbsC(N, n, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{n}{the number of between-sample random numbers}

\item{a}{the parameter of bivariate density}

\item{b}{the parameter of bivariate density}
}
\value{
a random sample of size n
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
set.seed(1211)
gibb_C<-gibbsC(120,15,1,1)
}
}
