\name{StatComp21093-package}
\alias{StatComp21093-package}
\alias{StatComp21093}
\docType{package}
\title{
  EM_mix and Gibbs_mix
}
\description{
  This package provides functions about EM Algorithm and Gibbs sampling for two-component Gaussian Mixture.
}
\details{
  R functions of EM algorithm and Gibbs sampling for two-component Gaussian Mixture and Rcpp function of chapter 9 homework are provided in this package.
}
\author{
Zhang Shan

Maintainer: Zhang Shan <zhangshan33@email.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
