% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{EM_mix}
\alias{EM_mix}
\title{EM_mix}
\usage{
EM_mix(mu1, sigma1, mu2, sigma2, p, N, rv)
}
\arguments{
\item{mu1}{the mean of Y1}

\item{sigma1}{the variance of Y1}

\item{mu2}{the mean of Y2}

\item{sigma2}{the variance of Y2}

\item{p}{the probability of the mixture distribution}

\item{N}{the number of samples}

\item{rv}{the random number from mixed normal distribution}
}
\value{
the maximum likelihood estimation of the mean, variance and the mixed probablity
}
\description{
EM Algorithm for two-component Gaussian Mixture
}
\examples{
\dontrun{
rv_mix<-function(N,p){
u1<-rnorm(N,0,1)
u2<-rnorm(N,2,1)
j<-runif(N)
k<-as.integer(j>p)
u<-k*u1+(1-k)u2
return (u)
}

p<-0.6
N<-1000
set.seed(2335)
rv<-rv_mix(N,p)
p_0<-0.5
mu1<-0.5
mu2<-1.5
sigma1<-sigma2<-sd(rv)
EM_mix(mu1,sigma1,mu2,sigma2,p_0,N,rv)
}
}
